+++ cli.go
@@ -1,6 +1,7 @@
 package cli
 
 import (
+	"os/exec"
 	"strings"
 	"time"
 
@@ -35,6 +36,36 @@ func Version(v string) {
 	app.Version = v
 }
 
+// Shell is a shortcut for exec.Command that configures the shell command to
+// use the CLI's Stdout and Stderr functions for output. It returns an
+// *exec.Cmd and you should run it manually.
+// If you want to execute the command and wait for it to finish, use Exec or
+// ExecAt instead for convenience.
+func Shell(name string, args ...string) *exec.Cmd {
+	return app.Shell("", name, args...)
+}
+
+// Exec executes a shell command with the given name and arguments. It returns
+// an error if the command fails to execute. The standard output and standard
+// error of the executed command are redirected to the CLI's Stdout and Stderr
+// functions, respectively. This allows you to capture and handle the output of
+// the executed command within your CLI application.
+func Exec(name string, args ...string) error {
+	return app.Exec("", name, args...)
+}
+
+// ExecAt executes a shell command with the given name and arguments in the
+// specified directory. It returns an error if the command fails to execute.
+// The standard output and standard error of the executed command are
+// redirected to the CLI's Stdout and Stderr functions, respectively. This
+// allows you to capture and handle the output of the executed command within
+// your CLI application. The dir parameter specifies the working directory for
+// the command. If dir is an empty string, the command will be executed in the
+// current working directory.
+func ExecAt(dir string, name string, args ...string) error {
+	return app.Exec(dir, name, args...)
+}
+
 // StdoutWith allows you to specify a custom function for handling standard
 // output. This can be useful for redirecting output to a file, logging system,
 // or for testing purposes. It is used to print the help text.
@@ -49,17 +80,11 @@ func StderrWith(fn func(msg string, args ...any)) {
 	app.Stderr = fn
 }
 
-// Print
-func Print(format string, v ...any) { app.Print(format, v...) }
+// Log prints a formatted message using the stdout function.
+func Log(format string, v ...any) { app.Print(format, v...) }
 
-// Println
-func Println(format string, v ...any) { app.Println(format, v...) }
-
-// Error
-func Error(format string, v ...any) { app.Error(format, v...) }
-
-// Errorln
-func Errorln(format string, v ...any) { app.Errorln(format, v...) }
+// LogError prints a formatted error message using the stderr function.
+func LogError(format string, v ...any) { app.Error(format, v...) }
 
 // UsePanicInsteadOfExit configures the CLI to panic instead of exiting when
 // an error  occurs or when a command finishes execution. This can be useful